+++ server/models/Payments.js
@@ -6,43 +6,47 @@ class NewModel extends Model {
         return this.query(
             `update payments
             set category_id=${params.categoryId}, type="${params.type}", amount=${params.amount}, date="${params.date}"
-            where id=${params.id}`
+            where id=${params.id}
+            and user_id=${params.userId}`
         );
     }
 
-    getAll() {
+    getAll(params) {
         return this.query(
             `select p.id, p.user_id, p.type as paymentType, p.amount, p.date, c.name as categoryName, c.id as categoryId
             from payments as p
             join categories as c
-            on p.category_id = c.id`
+            on p.category_id = c.id
+            where p.user_id=${params.userId}`
         );
     }
 
-    getLast() {
+    getLast(params) {
         return this.query(
             `select p.id, p.user_id, p.type as paymentType, p.amount, p.date, c.name as categoryName, c.id as categoryId
             from payments as p
             join categories as c
             on p.category_id = c.id
             where p.date <= NOW()
+            and p.user_id=${params.userId}
             order by p.date desc`
         );
     }
 
     delete(params) {
         return this.query(
-            `delete from payments where id=${params.id}`
+            `delete from payments where id=${params.id} and user_id=${params.userId}`
         );
     }
 
-    getFuture() {
+    getFuture(params) {
         return this.query(
             `select p.id, p.user_id, p.type as paymentType, p.amount, p.date, c.name as categoryName, c.id as categoryId
             from payments as p
             join categories as c
             on p.category_id = c.id
             where p.date > NOW()
+            and p.user_id=${params.userId}
             order by p.date asc`
         );
     }
